feat(ui): add plotly dep and ability to add/remove data points
Add dependency on react-plotly and accompanying type declarations
to create plots. Similarly add ability to add/remove data points
within the plot on the sidebar.

The plot appears to render correctly but axis titles are not getting
added to the plot as anticipated and no errors are being thrown.
Will need to investigate further.